+++ tests/metrics/test_metric.py
@@ -6,6 +6,7 @@ import cloudpickle
 import numpy as np
 import pytest
 import torch
+from torch import nn
 
 from pytorch_lightning.metrics.metric import Metric, MetricCollection
 
@@ -209,6 +210,25 @@ def test_state_dict(tmpdir):
     assert metric.state_dict() == OrderedDict(x=0)
     metric.persistent(False)
     assert metric.state_dict() == OrderedDict()
+
+
+def test_child_metric_state_dict():
+    """ test that child metric states will be added to parent state dict """
+    class TestModule(nn.Module):
+        def __init__(self):
+            super().__init__()
+            self.metric = Dummy()
+            self.metric.add_state('a', torch.tensor(0), persistent=True)
+            self.metric.add_state('b', [], persistent=True)
+            self.metric.register_buffer('c', torch.tensor(0))
+
+    module = TestModule()
+    expected_state_dict = {
+        'metric.a': torch.tensor(0),
+        'metric.b': [],
+        'metric.c': torch.tensor(0)
+    }
+    assert module.state_dict() == expected_state_dict
 
 
 @pytest.mark.skipif(not torch.cuda.is_available(), reason="Test requires GPU.")